Use a separate config file for documentation routes
Move the documentation routes out of the main app.js and into a
separate routes file within the documentation folder. Also move out
the route examples into the documentation routes instead of the app.
