+++ app.js
@@ -8,17 +8,18 @@ const nunjucks = require('nunjucks');
 // Local dependencies
 const authentication = require('./middleware/authentication');
 const automaticRouting = require('./middleware/auto-routing');
-const locals = require('./middleware/locals');
-const routes = require('./app/routes');
 const config = require('./app/config');
+const locals = require('./app/locals');
+const routes = require('./app/routes');
+const documentationRoutes = require('./docs/documentation_routes');
 
 // Set configuration variables
 const port = process.env.PORT || config.port;
 const useDocumentation = process.env.SHOW_DOCS || config.useDocumentation;
 
 // Initialise applications
-const app = express()
-const documentationApp = express()
+const app = express();
+const documentationApp = express();
 
 // Authentication middleware
 app.use(authentication);
@@ -35,7 +36,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 app.use('/nhsuk-frontend', express.static(path.join(__dirname, 'node_modules/nhsuk-frontend/packages')));
 app.use('/nhsuk-frontend', express.static(path.join(__dirname, 'node_modules/nhsuk-frontend/dist')));
 
-// Nunjucks configuration for app
+// Nunjucks configuration for application
 var appViews = [
   path.join(__dirname, 'app/views/'),
   path.join(__dirname, 'node_modules/nhsuk-frontend/packages/components')
@@ -58,7 +59,7 @@ if (useDocumentation) {
   // Documentation routes
   app.use('/docs', documentationApp);
 
-  // Nunjucks configuration for docs
+  // Nunjucks configuration for documentation
   var docViews = [
     path.join(__dirname, 'docs/views/'),
     path.join(__dirname, 'node_modules/nhsuk-frontend/packages/components')
@@ -68,16 +69,16 @@ if (useDocumentation) {
     autoescape: true,
     express: documentationApp
   });
-
-  // Docs router
-  documentationApp.get('/', function(req , res){ 
-    res.render('index');
-  });
   
-  // Automatically route docs pages
+  // Custom documentation routes
+  documentationApp.use('/', documentationRoutes);
+
+  // Automatically route documentation pages
   documentationApp.get(/^([^.]+)$/, function (req, res, next) {
     automaticRouting.matchRoutes(req, res, next)
   })
+
+  
 }
 
 // Run the application
