+++ docs/views/examples/branching/index.html
@@ -1,0 +1,5 @@
+{% extends 'layout.html' %}
+
+{% block content %}
+  
+{% endblock %}
