+++ docs/views/examples/passing-data/index.html
@@ -1,0 +1,5 @@
+{% extends 'layout.html' %}
+
+{% block content %}
+  {{ message }}
+{% endblock %}
